requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
drf-yasg>=1.21.7
PyJWT>=2.8.0
psycopg2-binary>=2.9.7
pytest-django>=4.5
